Initialize a new project database

Initialize a new project database

Usage

new_db(
  db_path = file.path(here::here(), ".sqlite"),
  mapping_fn = NULL,
  mapping_id = "ID",
  mapping_taxon = "Taxon",
  assemble_cpus = 6,
  assemble_memory = 16,
  seeds_db = glue::glue("/ref_dbs/getOrganelle/seeds/fish_mito.fasta"),
  labels_db = glue::glue("/ref_dbs/getOrganelle/labels/fish_mito.fasta"),
  getOrganelle =
    "-F 'anonym' -R 10 -k '21,45,65,85,105,115' --larger-auto-ws --expected-max-size 20000 --target-genome-size 16500",
  annotate_cpus = 6,
  annotate_memory = 16,
  annotate_ref_db = "Chordata",
  annotate_ref_dir = "/ref_dbs/Mitos2",
  mitos_opts = "--intron 0 --oril 0 --trna 0",
  trnaScan_opts = "-M vert",
  curate_cpus = 4,
  curate_memory = 8,
  curate_target = "fish_mito",
  curate_params = NULL
)

Arguments

db_path

Path to the new database file

mapping_fn

Path to the mapping file

mapping_id

Column name of the mapping file to use as the primary key

mapping_taxon

Column name of the mapping file contaning a Taxonomic identifier (eg, species name)

assemble_cpus

Default # cpus for assembly

assemble_memory

default memory (GB) for assembly

seeds_db

Path to the gotOrganelle seeds database

labels_db

Path to the gotOrganelle labels database

getOrganelle

Default getOrganelle command line options

annotate_cpus

Default # cpus for annotation

annotate_memory

Default memory (GB) for annotation

annotate_ref_db

Default Mitos2 reference database

annotate_ref_dir

Default Mitos2 reference database directory

mitos_opts

Default MITOS2 command line options

trnaScan_opts

Default tRNAscan-SE command line options

curate_cpus

Default # cpus for curation

curate_memory

Default memory (GB) for curation

curate_target

Default target database for curation

curate_params

Default curation parameters